+++ src/test/java/org/mariadb/jdbc/integration/codec/TinyIntCodecTest.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 // Copyright (c) 2012-2014 Monty Program Ab
 // Copyright (c) 2015-2023 MariaDB Corporation Ab
-
 package org.mariadb.jdbc.integration.codec;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -1027,7 +1026,7 @@ public class TinyIntCodecTest extends CommonCodecTest {
 
     assertTrue(rs.next());
     assertEquals(2, rs.getByte(2));
-    rs.updateObject(2, (Byte) null);
+    rs.updateObject(2, null);
     rs.updateRow();
     assertEquals(0, rs.getByte(2));
     assertTrue(rs.wasNull());
